Atom style callbacks pass indexes rather than an extension of Atom
The [TTFD](https://ttfd.vib.be) application uses this to display
selected atoms from a tracer metabolite. When the molecule itself is
used to store "selected" state the same instance must be used for
display. This prevents generalised React components where different
Molecule instances (of the same molecule) are used for display.